ended section 13

+++ src/app/api/todos/route.ts
@@ -1,7 +1,9 @@
 import prisma from '@/lib/prisma'
+import { getServerSession } from 'next-auth';
 import { NextResponse, NextRequest } from 'next/server'
 import { uptime } from 'process';
 import * as yup from 'yup';
+import { authOptions } from '../auth/[...nextauth]/route';
 
 export async function GET(request: Request) {
     const { searchParams } = new URL(request.url);
@@ -31,20 +33,32 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
     try {
+        const session = await getServerSession(authOptions);
+    
+
         const {complete, description} = await postSchema.validate(await request.json());
 
-        const todo = await prisma.todo.create({data:{complete, description}});
+        if( !session?.user) {
+            return NextResponse.json('No autorizado', {status:401})
+        }
+
+        const todo = await prisma.todo.create({data:{complete, description, userId: session.user.id}});
     
         return NextResponse.json(todo, { status: 201 });
     } catch (error) {
+        console.log(error)
         return NextResponse.json(error,{status:400});
     }
 }
 
 export async function DELETE(request: Request) {
     try {
+        const session = await getServerSession(authOptions);
+        if( !session?.user) {
+            return NextResponse.json('No autorizado', {status:401})
+        }
 
-        await prisma.todo.deleteMany({where: {complete: true}})
+        await prisma.todo.deleteMany({where: {complete: true, userId: session.user.id}})
 
         return NextResponse.json({messages: 'Tareas completadas borradas'}, { status: 200 });
     } catch (error) {
